movie_review_project/requirements.txt
django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-spectacular==0.27.2
gunicorn==22.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.9 
python-dotenv==1.0.1
requests==2.32.3